deletion working :)

+++ service/reservationService.js
@@ -38,22 +38,25 @@ async function updateReservation(accountId, restaurantId, newPartySize, newDate,
 }
 
 
-async function deleteReservation(accountId, restaurantId){
+async function deleteReservationFromDb(accountId, restaurantId) {
     return await appService.withOracleDB(async (connection) => {
         const result = await connection.execute(
-            `
-            DELETE FROM Reserve
-            WHERE accountId=` + accountId + ` AND restaurantId=` + restaurantId `;`
+            `DELETE FROM Reserve WHERE accountId = :accountId AND restaurantId = :restaurantId`,
+            [accountId, restaurantId], // Use bind parameters
+            { autoCommit: true }
         );
-        return result.rowsAffected && result.rowsAffected > 0; //Celine is not too sure but whatever! 
-    }).catch(() => {
-        return false;
-    });
+        console.log("Received accountId:", accountId, "Received restaurantId:", restaurantId);
+        return result.rowsAffected && result.rowsAffected > 0;
+        
+    }).catch (() => {
+        console.error('Error deleting reservation:', error);
+        return false; // Return false in case of an error
+    })
 }
 
 
 module.exports = {
     insertReservation,
     updateReservation,
-    deleteReservation,
+    deleteReservationFromDb,
 }
